requirements.txt
# requirements.txt
Flask
google-cloud-aiplatform
google-cloud-firestore